Extraido todos los eventos en una sola clase
