+++ Genetic Circuit/src/controlador/Controlador.java
@@ -1,19 +1,7 @@
 package controlador;
 
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.event.ItemEvent;
-import java.awt.event.ItemListener;
-import java.util.HashMap;
 import java.util.Timer;
 import java.util.TimerTask;
-import javax.swing.JButton;
-import javax.swing.JCheckBox;
-import javax.swing.JSpinner;
-import javax.swing.event.ChangeEvent;
-import javax.swing.event.ChangeListener;
-import javax.swing.JComboBox;
-
 import modelo.Modelo;
 import modelo.circuito.Circuito;
 import modelo.entidades.Entidad;
@@ -67,6 +55,8 @@ public class Controlador {
 	 */
 	private int tiempoVidaCache;
 	
+	private ControladorEventos controladorEventos;
+	
 	/**
 	 * Constructor que inicializa el controlador en el modo especificado. Si se le pasa "NORMAL",
 	 * se ejecuta el programa con normalidad, pero si se le pasa "EDITOR", su funcionalidad es guardar
@@ -98,29 +88,44 @@ public class Controlador {
 		 * o procede con la ejecución normal iniciando el circuito por defecto
 		 */
 		Ventana ventana = vista.getVentana();
-		ejecutarModoEditorCircuitos(ventana);
-		//Se carga el circuito por defecto al inicio del programa establecido en una constante
-		iniciarCircuito(CIRCUITO_INICIAL ,vista.getVentana());
+		/* Si se indica el modo de ejecución "EDITOR", al inicio del programa, llama al gestor de
+		 * circuitos para que ejecute el modo editor de circuitos, que directamente terminará
+		 * la ejecución prematuramente al guardar el circuito. Si el modo de ejecución es "NORMAL",
+		 * procede con la ejecución estándar del programa cargando un circuito en la ventana
+		 */
+		if(modoEjecucion.equals("EDITOR")) {
+			ejecutarModoEditorCircuitos(ventana);
+		} else {
+			//Se carga el circuito por defecto al inicio del programa establecido en una constante
+			iniciarCircuito(CIRCUITO_INICIAL ,vista.getVentana());
+		}
+		//Inicia los controladores encargados de manejar áreas más concretas
+		iniciarSubcontroladores();
 	}
 
+	/**
+	 * Inicia los controladores contenidos bajo el controlador principal para modularizar el 
+	 * comportamiento del programa
+	 */
+	private void iniciarSubcontroladores() {
+		controladorEventos = new ControladorEventos(this, modelo, vista);
+		vista.getPanelControl().asignarEventos();
+	}
+	
 	/** 
-	 * Si se indica el modo de ejecución "EDITOR", al inicio del programa
-	 * guardará como fichero en el proyecto el último circuito que hayamos diseñado
+	 * Guarda como fichero en el proyecto el último circuito que hayamos diseñado
 	 * para que se pueda seleccionar y cargar la siguiente vez que ejecute el programa.
-	 * El programa no continúa haciendo nada y termina salvo que cambiemos el valor de 
-	 * esa flag, pensada solo para el desarrollador y no el usuario
+	 * El programa no continúa haciendo nada y termina 
 	 * @param ventana gráfica necesaria para poder crear el circuito
 	 */
 	private void ejecutarModoEditorCircuitos(Ventana ventana) {
-		if(modoEjecucion.equals("EDITOR")) {
-			/* Se le tiene que pasa el nombre del circuito con el que se nombrará el fichero y
-			 * la ventana para poder colocar elementos relativos a ésta
-			 */
-			Circuito.guardarCircuito("circuito3", ventana);
-			//Cierra de manera segura la ventana gráfica y termina la ejecución del programa
-			vista.getVentana().exit();
-			System.exit(0);
-		}
+		/* Se le tiene que pasa el nombre del circuito con el que se nombrará el fichero y
+		 * la ventana para poder colocar elementos relativos a ésta
+		 */
+		Circuito.guardarCircuito("circuito3", ventana);
+		//Cierra de manera segura la ventana gráfica y termina la ejecución del programa
+		vista.getVentana().exit();
+		System.exit(0);
 	}
 	
 	/** 
@@ -297,7 +302,7 @@ public class Controlador {
 	/**
 	 * Vacía los datos del panel de control correspondientes a la última generación
 	 */
-	private void limpiarUltimaGeneracion() {
+	void limpiarUltimaGeneracion() {
 		modelo.getPoblacion().setEntidadMonitorizada(null); 
 		actualizarPanel("TiempoRecordActual", 0);
 		actualizarPanel("MejorAptitudActual", 0);
@@ -370,7 +375,7 @@ public class Controlador {
 	/**
 	 * Vacía la entidad monitorizada y sus datos del panel de control cuando su generación ya no exista
 	 */
-	private void limpiarEntidadMonitorizada() {
+	void limpiarEntidadMonitorizada() {
 		modelo.getPoblacion().setEntidadMonitorizada(null); 
 		actualizarPanel("Entidad", "-");
 		actualizarPanel("Posicion", "(0, 0)");
@@ -381,289 +386,6 @@ public class Controlador {
 		actualizarPanel("TiempoEntidad", 0); 
 		actualizarPanel("AptitudEntidad", 0);
 		actualizarPanel("EstadoEntidad", "-");
-	}
-	
-	/**
-	 * Iniciará la población de entidades a partir de los parámetros introducidos en el 
-	 * panel de control para que de comienzo al proceso evolutivo. Una vez iniciado 
-	 * cambia el botón de "Empezar" a un botón de "Siguiente" y cambiará su comportamiento. 
-	 * @author Alberto
-	 */
-	public class BtnEmpezarListener implements ActionListener {
-		@Override
-		public void actionPerformed(ActionEvent e) {
-			//Actualiza el estado a "realizando ciclo" ya que se empieza a ejecutar el proceso evolutivo
-			estado = Estado.REALIZANDO_CICLO;
-			//Iniciamos la población con los parámetros iniciales, incluido el punto de spawn del circuito
-			modelo.setPoblacionEntidades(setupPoblacion(), modelo.getCircuito().setSpawn());
-			/* Actualizamos la flag de parado para que la ventana pueda empezar a llamar a la función de
-			 * manipularProceso() cada frame
-			 */
-			parado = false; 
-			/* Cambiamos el texto del botón a "Siguiente", y le cambiamos el action listener por otro para
-			 * que se dispare un evento distinto cuando se interactúe con él
-			 */
-			//Guardamos el tiempo de vida en la cache para que pueda iniciar el proceso
-			tiempoVidaCache = modelo.getPoblacion().getTiempoVida();
-			PanelControl panelControl = vista.getPanelControl();
-			JButton btnProceder = panelControl.getBtnProceder();
-			btnProceder.setText("Siguiente");
-			btnProceder.addActionListener(new BtnSiguienteListener()); //añadimos el otro listener
-			btnProceder.removeActionListener(this); //eliminamos el actual
-			//Deshabilitamos el botón hasta que se pueda avanzar de generación
-			btnProceder.setEnabled(false); 
-			/* Deshabilita el selector de circuitos hasta que se reinicie al estado inicial
-			 * ya que cambiar el circuito en mitad del proceso es contraproducente
-			 */
-			panelControl.getcBoxCircuito().setEnabled(false);
-			//Habilita también los demás botones que ahora tendrán sentido utilizar
-			panelControl.getBtnReiniciar().setEnabled(true);
-			panelControl.getBtnPausar().setEnabled(true);
-			
-		}
-		
-		/**
-		 * Establece los parámetros iniciales del proceso evolutivo de la población a partir de
-		 * los valores introducidos en el panel de control
-		 * @return el mapa con los parámetros 
-		 */
-		private HashMap<String, Integer> setupPoblacion() {
-			PanelControl panelControl = vista.getPanelControl();
-			HashMap<String, Integer> poblacionParams = new HashMap<String, Integer>();
-			poblacionParams.put("NumEntidades", panelControl.getTotalPoblacion());
-			poblacionParams.put("TasaMutacion", panelControl.getTasaMutacion());	
-			poblacionParams.put("TiempoVida", panelControl.getTiempoVida()); 
-			poblacionParams.put("TiempoObjetivo", panelControl.getTiempoObjetivo());
-			return poblacionParams;
-		}
-	}
-	
-	/**
-	 * Actualiza la flag de parado a false para que pueda realizar el ciclo de vida de las entidades 
-	 * @author Alberto
-	 */
-	public class BtnSiguienteListener implements ActionListener {
-		@Override
-		public void actionPerformed(ActionEvent e) {
-			//Actualiza el estado a "realizando ciclo" ya que se reanuda el proceso evolutivo
-			estado = Estado.REALIZANDO_CICLO;
-			parado = false;
-			//Habilita el botón de pausa ya que ya no provocaría que iniciase otro ciclo
-			vista.getPanelControl().getBtnPausar().setEnabled(true);
-			//Desactiva el propio botón hasta que se pueda pasar a la siguiente generación
-			vista.getPanelControl().getBtnProceder().setEnabled(false);
-			limpiarEntidadMonitorizada();
-			limpiarUltimaGeneracion();
-		}
-	}
-	
-	/**
-	 * Actualiza la flag de parado a true para que pause el ciclo de vida de las entidades
-	 * hasta reanudar pulsando el mismo botón (actualizando la flag de nuevo)
-	 * @author Alberto
-	 */
-	public class BtnPausarListener implements ActionListener {
-		@Override
-		public void actionPerformed(ActionEvent e) {
-			parado = !parado; //Invierte el estado de la flag
-			PanelControl panelControl = vista.getPanelControl();
-			//Actualiza el texto a "Reanudar" o "pausar" según está parado o no
-			String textoBtn = (parado) ? "Reanudar" : "Pausar";
-			panelControl.getBtnPausar().setText(textoBtn);
-			/* Si se ha pausado la ejecución, actualiza el estado a "pausado", y si se ha
-			 * reanudado, se actualiza a "realizando ciclo */
-			estado = (parado) ? Estado.PAUSADO : Estado.REALIZANDO_CICLO;
-		}
-	}
-	
-	/**
-	 * Para el actual proceso evolutivo y restaura el panel de control al estado inicial
-	 * del programa para poder iniciar de nuevo otra ejecución de cero
-	 * @author Alberto
-	 */
-	public class BtnReiniciarListener implements ActionListener {
-		@Override
-		public void actionPerformed(ActionEvent e) {
-			//Actualiza el estado a "en espera" ya que se ha dejado de ejecutar el proceso evolutivo
-			estado = Estado.EN_ESPERA;
-			//Impide que la ventana siga actualizando hasta que volvamos a empezar
-			parado = true; 
-			Ventana ventana = vista.getVentana();
-			/* Hay que reiniciar el número de frames porque si no se produce un bug
-			 * en el que tras empezar de nuevo, empieza a contar desde el frame en el
-			 * que estaba al pausar el proceso antes de reiniciar, rompiendo el programa
-			 */
-			ventana.setNumFramesGen(0); 
-			/* Dejamos activado sólo el botón de empezar y el selector de circuitos, que
-			 * ahora no tendría problema para poder cargar otro distinto al haber empezado
-			 * otro proceso
-			 */
-			PanelControl panelControl = vista.getPanelControl();
-			panelControl.getcBoxCircuito().setEnabled(true);
-			panelControl.getBtnReiniciar().setEnabled(false);
-			panelControl.getBtnPausar().setEnabled(false);
-			panelControl.getBtnPausar().setText("Pausar");
-			JButton btnProceder = panelControl.getBtnProceder();
-			btnProceder.setEnabled(true);
-			/* Cambiamos el botón de "Siguiente" por "Empezar" y sustituimos su listener
-			 * por el que le corresponde ahora
-			 */
-			btnProceder.setText("Empezar");
-			btnProceder.removeActionListener(btnProceder.getActionListeners()[0]);
-			btnProceder.addActionListener(new BtnEmpezarListener());
-			//Reseteamos la información mostrada en el panel de control
-			reiniciarInfoPanel(panelControl);
-		}
-		
-		/* Reinicia todos los valores de información del panel de control sobre el 
-		 * anterior proceso a cero pero deja los parámetros editables sin alterar
-		 * por si quiere repetir un proceso igual que antes
-		 */
-		private void reiniciarInfoPanel(PanelControl panelControl) {
-			panelControl.setValor("TiempoRecord", 0);
-			panelControl.setValor("MejorAptitud", 0);
-			panelControl.setValor("Metas", 0);
-			panelControl.setValor("Colisiones", 0);
-			panelControl.setValor("Generacion", 0);
-			limpiarEntidadMonitorizada();
-			limpiarUltimaGeneracion();
-		}
-	}
-	
-	/**
-	 * Cierra el programa sin importar el estado en el que se encuentre
-	 * @author Alberto
-	 */
-	public class BtnSalirListener implements ActionListener {
-		@Override
-		public void actionPerformed(ActionEvent e) {
-			vista.getVentana().exit();
-			System.exit(0);
-		}
-	}
-	
-	/**
-	 * Activa o desactiva según se marque la checkbox el modo automático. Si se activa
-	 * las generaciones se sucederán sólas sin necesidad del input del usuario para
-	 * continuar su ciclo de vida. Si se desactiva el usuario deberá pulsar el botón
-	 * de "Siguiente" para avanzar de generación.
-	 * @author Alberto
-	 */
-	public class CbModoAutoListener implements ActionListener {
-	
-		@Override
-		public void actionPerformed(ActionEvent e) {
-			//Obtenemos si el checkbox que lanza el evento está o no activado
-			JCheckBox cbModoAuto = (JCheckBox) e.getSource();
-			boolean activado = cbModoAuto.isSelected();
-			//Activa o desactiva el modo automático según esté seleccionado
-			modoAutomatico = activado;
-			
-			JButton btnProceder = vista.getPanelControl().getBtnProceder();
-			/* Comprueba si el botón proceder es ahora mismo "Siguiente" ya que no debe
-			 * modificarlo si está en modo "Empezar", ya que solo debe comenzar una población
-			 * de esa manera.
-			 */
-			if(btnProceder.getText().equals("Siguiente")) {
-				/* Si el botón "Siguiente" está habilitado, implica que la generación actual
-				 * ha terminado su ciclo y que está esperando a que se pulse para continuar.
-				 * Por tanto, nos interesa que cuando activemos el modo automático inicie
-				 * la siguiente generación directamente
-				 */
-				if(btnProceder.isEnabled()) {
-					parado = false;
-					/* Deshabilita el botón de "Siguiente" solo si el proceso estaba parado,
-					 * para que en el siguiente ciclo ya no permita avanzar manualmente
-					 */
-					btnProceder.setEnabled(!activado);
-				} 
-			}
-			
-		}
-	}
-	
-	/**
-	 * Establece un nuevo valor para el parámetro de la población de entidades, a través 
-	 * del cambio detectado en el spinner. Realiza las restricciones necesarias antes
-	 * de modificar el parámetro deseado. Sirve para cualquiera de los spinners
-	 * @author Alberto
-	 */
-	public class SpParamListener implements ChangeListener {
-		@Override
-		public void stateChanged(ChangeEvent e) {
-			JSpinner spinner = (JSpinner) e.getSource();
-			//Obtiene el nombre del parámetro a cambiar a partir del nombre del spinner
-			String param = spinner.getName(); 
-			int valor = (int) spinner.getValue(); //valor cambiado en el spinner
-			/* Si el spinner es el correspondiente al tiempo de vida, debe controlar que
-			 * su valor nunca sea menor o igual que el tiempo objetivo, así que en ese caso
-			 * altera el valor para que esté por encima
-			 */
-			if(param.equals("TiempoVida")) {
-				int valorSpObjetivo = (int)vista.getPanelControl().getSpTiempoObjetivo().getValue();
-				if(valor <= valorSpObjetivo) {
-					spinner.setValue(valorSpObjetivo + 10);
-					valor = (int)spinner.getValue();
-				}
-			}
-			/* Al principio del programa la población no está inicializada así que no debe actualizar
-			 * ningún parámetro en ese caso, ya se encarga el botón "Empezar" de capturar los valores
-			 */
-			if(modelo.getPoblacion() != null) {
-				//Si población ya está inicializada puede actualizarla sin problema
-				actualizarParam(param, valor);
-			}
-		}
-
-		/**
-		 * Según el nombre del parámetro actualiza el correspondiente de la población
-		 * con el valor nuevo
-		 * @param param 
-		 * @param valor
-		 */
-		private void actualizarParam(String param, int valor) {
-			switch(param) {
-				case "NumEntidades":
-					modelo.getPoblacion().setNumEntidades(valor);
-					break;
-				case "TasaMutacion":
-					modelo.getPoblacion().setTasaMutacion(valor);
-					break;
-				case "TiempoObjetivo":
-					modelo.getPoblacion().setTiempoObjetivo(valor);
-					break;
-				case "TiempoVida":
-					modelo.getPoblacion().setTiempoVida(valor);
-					break;
-			}
-		}
-	}
-	
-	/**
-	 * Cambia el circuito cargando el fichero correspondiente a la opción que se ha 
-	 * seleccionado en el combo box y reestableciendo los parámetros necesarios en los
-	 * elementos del proceso como la meta, obstáculos y la población a partir del circuito.
-	 * En el siguiente frame se actualizara el circuito mostrado en la ventana gráfica.
-	 * @author Alberto
-	 */
-	public class CboxCircuitoListener implements ItemListener {
-		
-		@Override
-		public void itemStateChanged(ItemEvent e) {
-			//Solo se activa si se selecciona una opción distinta a la actual
-			if(e.getStateChange() == ItemEvent.SELECTED) {
-				@SuppressWarnings("unchecked")
-				JComboBox<String> cBoxCircuito = (JComboBox<String>) e.getSource();
-				//Obtiene el nombre del circuito a partir del item seleccionado
-				String circuitoSeleccionado = (String) cBoxCircuito.getSelectedItem();
-				/* Inicia los elementos (meta y obstáculos) del circuito tras cargar el ficheros
-				 * con el nombre del circuito seleccionado
-				 */
-				iniciarCircuito(circuitoSeleccionado, vista.getVentana());
-			}
-			
-		}
-		
 	}
 	
 	public Vista getVista() {
@@ -684,6 +406,26 @@ public class Controlador {
 
 	public Estado getEstado() {
 		return estado;
+	}
+
+	public void setModoAutomatico(boolean modoAutomatico) {
+		this.modoAutomatico = modoAutomatico;
+	}
+
+	public void setParado(boolean parado) {
+		this.parado = parado;
+	}
+
+	public void setEstado(Estado estado) {
+		this.estado = estado;
+	}
+
+	public void setTiempoVidaCache(int tiempoVidaCache) {
+		this.tiempoVidaCache = tiempoVidaCache;
+	}
+
+	public ControladorEventos getControladorEventos() {
+		return controladorEventos;
 	}
 
 }
